refactor(rockpaperscissors): Refactored the logic to optimise for Ties and utilising nested ifs
Also added some further tests. Note: When following the tutorial, the author had to manually check the variables and run the code to ascertain if the code still worked. Turned out it didn't. Lucky for me, tests ensure that I knew my expected outcome and it was a piece of cake!